+++ st2api/tests/unit/controllers/v1/test_executions.py
@@ -26,6 +26,7 @@ except ImportError:
 from six.moves import filter
 
 from st2common.constants import action as action_constants
+from st2common.constants.secrets import MASKED_ATTRIBUTE_VALUE
 from st2common.models.db.execution import ActionExecutionDB
 from st2common.models.db.execution import ActionExecutionOutputDB
 from st2common.persistence.execution import ActionExecution
@@ -133,6 +134,14 @@ ACTION_4 = {
     }
 }
 
+ACTION_INQUIRY = {
+    'name': 'st2.dummy.ask',
+    'description': 'another test description',
+    'enabled': True,
+    'pack': 'wolfpack',
+    'runner_type': 'inquirer',
+}
+
 LIVE_ACTION_1 = {
     'action': 'sixpack.st2.dummy.action1',
     'parameters': {
@@ -162,6 +171,44 @@ LIVE_ACTION_3 = {
 
 LIVE_ACTION_4 = {
     'action': 'starterpack.st2.dummy.action4',
+}
+
+LIVE_ACTION_INQUIRY = {
+    'parameters': {
+        'route': 'developers',
+        'schema': {
+            'type': 'object',
+            'properties': {
+                'secondfactor': {
+                    'secret': True,
+                    'required': True,
+                    'type': u'string',
+                    'description': 'Please enter second factor for authenticating to "foo" service'
+                }
+            }
+        }
+    },
+    'action': 'wolfpack.st2.dummy.ask',
+    'result': {
+        'users': [],
+        'roles': [],
+        'route': 'developers',
+        'ttl': 1440,
+        'response': {
+            'secondfactor': 'supersecretvalue'
+        },
+        'schema': {
+            'type': 'object',
+            'properties': {
+                'secondfactor': {
+                    'secret': True,
+                    'required': True,
+                    'type': 'string',
+                    'description': 'Please enter second factor for authenticating to "foo" service'
+                }
+            }
+        }
+    }
 }
 
 FIXTURES_PACK = 'generic'
@@ -196,12 +243,17 @@ class ActionExecutionControllerTestCase(BaseActionExecutionControllerTestCase, F
         post_resp = cls.app.post_json('/v1/actions', cls.action4)
         cls.action4['id'] = post_resp.json['id']
 
+        cls.action_inquiry = copy.deepcopy(ACTION_INQUIRY)
+        post_resp = cls.app.post_json('/v1/actions', cls.action_inquiry)
+        cls.action_inquiry['id'] = post_resp.json['id']
+
     @classmethod
     def tearDownClass(cls):
         cls.app.delete('/v1/actions/%s' % cls.action1['id'])
         cls.app.delete('/v1/actions/%s' % cls.action2['id'])
         cls.app.delete('/v1/actions/%s' % cls.action3['id'])
         cls.app.delete('/v1/actions/%s' % cls.action4['id'])
+        cls.app.delete('/v1/actions/%s' % cls.action_inquiry['id'])
         super(BaseActionExecutionControllerTestCase, cls).tearDownClass()
 
     def test_get_one(self):
@@ -1051,6 +1103,31 @@ class ActionExecutionControllerTestCase(BaseActionExecutionControllerTestCase, F
         finally:
             action_constants.WORKFLOW_RUNNER_TYPES.remove(ACTION_1['runner_type'])
 
+    def test_get_inquiry_mask(self):
+        """Ensure Inquiry responses are masked when retrieved via ActionExecution GET
+
+        The reason this test is included here is so that we can verify that the ActionExecution
+        GET function properly masks fields within an Inquiry response.
+        TODO(mierdin): This test, and the constants it uses, should not be necessary here
+        once Inquiries get their own data model.
+        """
+
+        post_resp = self._do_post(LIVE_ACTION_INQUIRY)
+        actionexecution_id = self._get_actionexecution_id(post_resp)
+        get_resp = self._do_get_one(actionexecution_id)
+        self.assertEqual(get_resp.status_int, 200)
+
+        resp = json.loads(get_resp.body)
+        self.assertEqual(resp['result']['response']['secondfactor'], MASKED_ATTRIBUTE_VALUE)
+
+        post_resp = self._do_post(LIVE_ACTION_INQUIRY)
+        actionexecution_id = self._get_actionexecution_id(post_resp)
+        get_resp = self._do_get_one(actionexecution_id, params={'show_secrets': True})
+        self.assertEqual(get_resp.status_int, 200)
+
+        resp = json.loads(get_resp.body)
+        self.assertEqual(resp['result']['response']['secondfactor'], "supersecretvalue")
+
 
 class ActionExecutionOutputControllerTestCase(BaseActionExecutionControllerTestCase,
                                               FunctionalTest):
@@ -1081,7 +1158,6 @@ class ActionExecutionOutputControllerTestCase(BaseActionExecutionControllerTestC
             output_params['data'] = 'stdout mid 1\n'
             output_db = ActionExecutionOutputDB(**output_params)
             ActionExecutionOutput.add_or_update(output_db)
-            pass
 
         # Since the API endpoint is blocking (connection is kept open until action finishes), we
         # spawn an eventlet which eventually finishes the action.
@@ -1091,15 +1167,19 @@ class ActionExecutionOutputControllerTestCase(BaseActionExecutionControllerTestC
             output_db = ActionExecutionOutputDB(**output_params)
             ActionExecutionOutput.add_or_update(output_db)
 
+            eventlet.sleep(1.0)
+
             # Transition execution to completed state so the connection closes
             action_execution_db.status = action_constants.LIVEACTION_STATUS_SUCCEEDED
             action_execution_db = ActionExecution.add_or_update(action_execution_db)
 
         eventlet.spawn_after(0.2, insert_mock_data)
         eventlet.spawn_after(1.5, publish_action_finished, action_execution_db)
+
+        # Retrieve data while execution is running - endpoint return new data once it's available
+        # and block until the execution finishes
         resp = self.app.get('/v1/executions/%s/output' % (str(action_execution_db.id)),
                             expect_errors=False)
-
         self.assertEqual(resp.status_int, 200)
         lines = resp.text.strip().split('\n')
         self.assertEqual(len(lines), 3)