+++ st2common/st2common/runners/base.py
@@ -14,7 +14,6 @@
 # limitations under the License.
 
 import abc
-import pkgutil
 
 import six
 import yaml
@@ -27,7 +26,7 @@ from st2common.exceptions.actionrunner import ActionRunnerCreateError
 from st2common.util import action_db as action_utils
 from st2common.util.loader import register_runner, register_callback_module
 from st2common.util.api import get_full_public_api_url
-
+from st2common.util.deprecation import deprecated
 
 __all__ = [
     'ActionRunner',
@@ -50,12 +49,15 @@ def get_runner(module_name, config=None):
     """
 
     LOG.debug('Runner loading python module: %s', module_name)
+
     try:
         # TODO: Explore modifying this to support register_plugin
         module = register_runner(module_name)
     except Exception as e:
-        LOG.exception('Failed to import module %s.', module_name)
-        raise ActionRunnerCreateError(e)
+        msg = ('Failed to import runner module %s' % module_name)
+        LOG.exception(msg)
+
+        raise ActionRunnerCreateError('%s\n\n%s' % (msg, str(e)))
 
     LOG.debug('Instance of runner module: %s', module)
 
@@ -75,8 +77,14 @@ def get_metadata(package_name):
 
     :rtype: ``list`` of ``dict``
     """
-    file_path = pkgutil.get_data(package_name, 'metadata/runner.yaml')
-    metadata = yaml.safe_load(file_path)
+    import pkg_resources
+
+    file_path = pkg_resources.resource_filename(package_name, 'runner.yaml')
+
+    with open(file_path, 'r') as fp:
+        content = fp.read()
+
+    metadata = yaml.safe_load(content)
     return metadata
 
 
@@ -159,7 +167,11 @@ class ActionRunner(object):
                 result
             )
 
+    @deprecated
     def get_pack_name(self):
+        return self.get_pack_ref()
+
+    def get_pack_ref(self):
         """
         Retrieve pack name for the action which is being currently executed.
 
@@ -191,7 +203,7 @@ class ActionRunner(object):
         :rtype: ``dict``
         """
         result = {}
-        result['ST2_ACTION_PACK_NAME'] = self.get_pack_name()
+        result['ST2_ACTION_PACK_NAME'] = self.get_pack_ref()
         result['ST2_ACTION_EXECUTION_ID'] = str(self.execution_id)
         result['ST2_ACTION_API_URL'] = get_full_public_api_url()