+++ st2common/st2common/runners/python_action_wrapper.py
@@ -94,6 +94,13 @@ class ActionService(object):
         return self._datastore_service
 
     ##################################
+    # General methods
+    ##################################
+
+    def get_user_info(self):
+        return self.datastore_service.get_user_info()
+
+    ##################################
     # Methods for datastore management
     ##################################