+++ st2common/tests/unit/test_unit_testing_mocks.py
@@ -34,6 +34,11 @@ class MockSensorClass(object):
 
 class BaseMockResourceServiceTestCase(object):
     class TestCase(unittest2.TestCase):
+        def test_get_user_info(self):
+            result = self.mock_service.get_user_info()
+            self.assertEqual(result['username'], 'admin')
+            self.assertEqual(result['rbac']['roles'], ['admin'])
+
         def test_list_set_get_delete_values(self):
             # list_values, set_value
             result = self.mock_service.list_values()