+++ CHANGELOG.rst
@@ -15,12 +15,27 @@ Added
   ``error`` or higher (or similar). (new feature) #3824
 * Add stevedore related metadata to Python package setup.py files for runner packages. This way
   runners can be installed using pip and dynamically enumerated and loaded using stevedore and
-  corresponding helper functions. (new feature)
+  corresponding helper functions.
+
+  All runners are now also fully fledged Python packages (previously they were single module
+  Python packages which caused various install and distribution related issues when installing
+  them via pip) (new feature)
 * Add new ``search`` rule criteria comparison operator. For the usage, please refer to the
   documentation. (new feature) #3833
 
   Contributed by @ahubl-mz.
-  
+* Add new ``/api/v1/user`` API endpoint. This API endpoint is only available to the authenticated
+  users and returns various metadata on the authenticated user (which method did the user use to
+  authenticate, under which username the user is authenticated, which RBAC roles are assignment to
+  this user in case RBAC is enabled, etc.) (new feature) #3831
+* Add new ``get_user_info`` method to action and sensor service. With this method, user can
+  retrieve information about the user account which is used to perform datastore operations inside
+  the action and sensor service. (new feature) #3831
+* Add ability to share common code between python sensors and python actions. You can now place
+  common code inside a ``lib`` directory inside a pack (with an ``__init__.py`` inside ``lib``
+  directory to declare it a python package). You can then import the common code in sensors and
+  actions. Please refer to documentation for samples and guidelines. #3490
+
 Changed
 ~~~~~~~
 
@@ -44,6 +59,8 @@ Changed
   gitpython, pymongo, stevedore, paramiko, prompt-toolkit, flex). #3830
 * Update log messages in the datastore service to correctly use ``DEBUG`` log level instead of
   ``AUDIT``. #3845
+* Mask values in an Inquiry response displayed to the user that were marked as "secret" in the
+  inquiry's response schema. #3825
 * Added the ability of ``st2 key load`` to load keys from both JSON and YAML files.
   Files can now contain a single KeyValuePair, or an array of KeyValuePairs.
   Updated ``st2 key load`` to load all properties of a key/value pair, now
@@ -61,6 +78,8 @@ Fixed
   (bug fix) #3824
 * Fix ``st2 execution tail [last]`` CLI command so it doesn't throw an exception if there are no
   executions in the database. (bug fix) #3760 #3802
+* Fix 'NameError: name 'cmd' is not defined' error when using ``linux.service`` with CentOS systems.
+  #3843. Contributed by @shkadov
 
 2.5.0 - October 25, 2017
 ------------------------